Port Ice Cream Calculator library from Apps Script
This Apps Script library was developed for an Ice Cream Calculator
in a Google Sheets spreadsheet. It is a very poor javascript/typescript
library, likely so unergonomic as to be non-functional for the purposes
of this project. However, the underlying math is functional, and it's
decently tested, so it's being used as a jumping-off point. The port
introduced as few modifications as possible to make it valid typescript.

Ported from:

Version 46 on 22 Mar 2025, 22:20
Add support for computeAndDisplayFruitMedley
AKfycbzXSXjEUPkh9LbZEDmA3TC-IkuMgZcx5KkE3dQlPtZA1NYXeXFGoVCwHtd8y17KZ3vp
https://script.google.com/macros/library/d/13qgahGqZwn9QI8KeplgDyqRn4x2WQZFqNZDmu2nF2uHTsZ7OCcYc0aru/46