Make every `IngredientRow` an independent state
My making every `IngredientRow` an independent `IngredientRowState` we
can check every time a row changes if the name is a valid ingredient
name, from a list that is only fetched once. Then, only if we detect a
valid name do we do a fetch for the ingredient information. This is much
more performant, even without a cache and with simulated latencies.