Add skeleton for supporting "wasm" and "backend"
`wasm-bindgen` and `diesel` dependencies cannot be included together in
a build of the `sci-cream` Rust crate because `diesel` is not supported
for WASM targets. So we need to separate the builds by features, none
for just the core library to be used in any application, "wasm" for WASM
targets to be used client-side, and "backend" to include support for
database management with `diesel`.