Add some improvements to the GridComp border issue
Remove the duplicate border at the top and left-most column of the
header rows, but there is still a weird blank line on the left-most,
and the right-most still has the too-thick problem.

It's possible to use `enabledHeaderStringKeysIndexed` and check for
whether `index === 0` to control color of the left-most border, but
it's messy and there must be a better solution to this.