Remove all Option<_> from composition types
Although all `None` values are implied to be zero, there is a small
distinction between a value not being specified and being specified
as zero, particularly with regards to displaying compositions. However,
the distinction is not worth the ergonomics cost associated with all
the `Option<_>`s everywhere, so remove all `Option<_>` from composition
types and instead explicitly set `0f64`/`f64::default()` on all values.

Implemenet the following generic function to iterate over struct fields
while downcasting to an expected type, returning an `Iterator`:
  `util::iter_fields_as<T, I:Iterable>(...) -> impl Iterator<Item = T>`

Note that after this change `packages/app/.../IngredientCompositionGrid`
will display a lot of zeroes, requiring a follow-up change to display
zeroes as blanks.