Add support for Auto/Custom IngCompGrid columns
Add support for `ColumnFilter.Auto` and `.Custom`, where `Auto` displays
all the columns that have at least one non-zero value, and `Custom`
supports a manual selection of the columns to be displayed, which is
kept in a client state. Implement functionality for checkboxes to update
the state of selected columns.