Add WASM constructor and to_pod/pac_js to `Sugars`
These serve as a proof-of-concept to make the Rust types constructible
and testable from the exported WASM bindings. Adopt the conventions
that, when the Rust signature is not supported by `wasm-bindgen`, e.g.
when returning Result<...>, a similarly named functions with the `_js`
suffix will be provided, with a signature supported by `wasm-bindgen`.

Add `sci-cream/src/ts/composition.test.ts` with unit tests for the
construction of `Sugars` in typescript, and the to_pod/pac_js functions.