Remove Sugars.unspecified field, to_pod/pac -> f64
Supporting the `unspecified` field adds a possible error variant that
requires a lot of checking. It's currently unclear if that field would
be of any use, compared to a best guess of the sugars composition. As
such, given the cost, remove the field and add it only if proven
necessary in the future.