Refactor Sugars/SweetsSpec explicit dry/total
Refactor `SugarsSpec` and `SweetenersSpec` into a unified
`SweetenersSpec` with a `basis` field to specify if composition values
are given as a percentage of the dry weight (solids) or of the total
weight. This makes specifications clearer, less error prone, allows for
better error checking, and allows more flexibility to specify
compositions as given by sources, without modifications.