Add #[serde(deny_unknown_fields)] to spec structs
Many of the spec and composition structs use #[serde(default)] to allow
not specifying fields that are usually zero. This makes specifying
ingredients more convenient, but it causes typos to not generate errors,
which can be problematic. As such, add #[serde(deny_unknown_fields)] to
all these structs so that typos produce errors when deserializing.

@note This catches a previous typo in sweeteners.json, where
"polysaccharides" was mistyped as "polysaccharide", which is fixed
in this commit.