Refactor PropKey to nest CompKey and new FpdKey
Instead of duplicating all the variants form CompKey into PropKey, make
PropKey contain a variant with a CompKey value, and also a variant with
a value of the new FpdKey. This greatly simplifies the code and
maintenance burden.

However, this new PropKey is not supported in wasm (enums with values
are not supported), so re-create the functionality in TypeScript via a
union enum: `const PropKeyObj = { ...CompKey, ...FpdKey} as const` and a
few utilities to facilitate its use:

   `isCompKey(PropKey)`
   `isFpdKey(PropKey)`
   `getPropKeys`
   `prop_key_as_med_str_js(PropKey)`
   `getMixProperty(MixProperties, PropKey)`

Use these new types and utilities in `packages/app` components.