Add `other_solids` to `SweetenersSpec`, fix bug
Add `other_solids` field to `SweetenersSpec` to allow specifying
impurities that may be found in some ingredients, e.g. 1% in Honey.
The previous 1% polysaccharides can now be properly specified as SNFS.
This field is optional and should rarely be needed.

Fix a bug where `polysaccharides` were being counted as
Solids.sweeteners, instead of the correct .snfs.