Use font-mono and fixed decimal point format
Refactor formatting of composition values to use a monospace font, and
to left and right pad values so that the decimal point always has a
fixed position. Do not display decimal trailing zeroes. For values >= 10
display one decimal digit, and for values < 10 display 2 decimal digits.
For values >= 1000, display as (val/1000).toFixed(1)k.