Refactor recipe state to fix React Hook lint warns
Refactor the entirety of the recipe states, how they're updated, the use
of useEffect, etc. to fix React Hook lint warnings about using useState
in a loop or callback. Now there is a single state with all the recipes,
which are updated in RecipeGrid, and just the data is passed down to
other components.