Fix RecipeGrid field display delay when paste
When pasting a recipe into a RecipeGrid, displaying the pasted names and
quantities is delayed waiting for the Ingredient to be feteched, which
looks jarring. Fix this issue by immediately calling `updateRecipe` with
the new name and quantity, but keeping the existing Ingredient, so that
the name/quantity is immediately displayed. Then, only if necessary,
fetch a new Ingredient and call `updateRecipe` again, so that only the
computations are delayed waiting for the fetch, and only if we don't
already have the correct Ingredient.