Use fixed num of col in ReactGridLayout & reactive
Change back to using a fixed number of columns with react-grid-layout.
The prvious implementation of changing the number of clumns based on the
width was causing issues; this is the way it's meant to be used. This is
not without its problems; some components don't look good when they
resize too wide, e.g. RecipeGrid and MixPropertiesGrid, but we can deal
with that later. For now, allow all component to resize dynamically, at
least horizontally. Change the number of ingredient rows from 21 to 22
to accomodate these changes and some better margin.