Prevent stale ingredient rows during pre-fetch
If a recipe is pasted quickly whilst the page is still loading, then
validIngredients/ingrecientCache may not yet be populated, which will
cause the ingredients to not be updated (they appear not valid) and so
the dependent component to not be updated. Adding a useEffect dependent
on validIngredients.length and ingredientCache.size seems to fix this
problem, although it causes 80 calls to updateIngRow on mount; it's
unclear why that is the case, and why 80; 20 rows * 3 recipes = 60.
Looks like each recipe being rendered gets refreshed twice on mount;
need to investigate.