Change grid-layout cols to 24, and set correct y
Using 24 instead of 12 columns allows for more flexibility on the
upcoming support to add dynamic width contraints to some components,
e.g. RecipeGrid and MixPropertiesGrid. Also, setting 'y' values that
correctly place the components based on REACT_GRID_COMPONENT_HEIGHT
allows a `horizontalCompactor` to be used with the expected effect.